+++ src/main/java/de/geheimagentnr1/world_pre_generator/config/GenerationType.java
@@ -1,0 +1,7 @@
+package de.geheimagentnr1.world_pre_generator.config;
+
+public enum GenerationType {
+	
+	SERIAL,
+	SEMI_PARALLEL
+}