+++ src/main/java/de/geheimagentnr1/world_pre_generator/config/ServerConfig.java
@@ -15,6 +15,18 @@ public class ServerConfig extends AbstractConfig {
 	private static final String SEND_FEEDBACK_KEY = "send_feedback";
 	
 	@NotNull
+	private static final String GENERATION_KEY = "generation";
+	
+	@NotNull
+	private static final List<String> GENERATION_TYPE_KEY = List.of( GENERATION_KEY, "type" );
+	
+	@NotNull
+	private static final List<String> GENERATION_SEMI_PARALLEL_TASK_COUNT_KEY = List.of(
+		GENERATION_KEY,
+		"semi_parallel_task_count"
+	);
+	
+	@NotNull
 	private static final String DELAYS_KEY = "delays";
 	
 	@NotNull
@@ -22,15 +34,6 @@ public class ServerConfig extends AbstractConfig {
 	
 	@NotNull
 	private static final List<String> SAVE_DELAY_KEY = List.of( DELAYS_KEY, "save" );
-	
-	@NotNull
-	private static final String PARALLEL_KEY = "parallel";
-	
-	@NotNull
-	private static final List<String> PARALLEL_ENABLED_KEY = List.of( PARALLEL_KEY, "enabled" );
-	
-	@NotNull
-	private static final List<String> PARALLEL_COUNT_KEY = List.of( PARALLEL_KEY, "count" );
 	
 	public ServerConfig( @NotNull AbstractMod _abstractMod ) {
 		
@@ -53,7 +56,25 @@ public class ServerConfig extends AbstractConfig {
 	@Override
 	protected void registerConfigValues() {
 		
-		registerConfigValue( "Shall a pregeneration feedback send to all online players?", SEND_FEEDBACK_KEY, true );
+		registerConfigValue( "Shall a pre generation feedback send to all online players?", SEND_FEEDBACK_KEY, true );
+		push( "Parameters for generation", GENERATION_KEY );
+		registerConfigValue(
+			List.of(
+				"Type of generation",
+				GenerationType.SERIAL + ": Every chunk one after another is generated.",
+				GenerationType.SEMI_PARALLEL + ": Multiple chunk generation tasks are created, " +
+					"but the generation is still serial (speed up the chunk generation)."
+			),
+			GENERATION_TYPE_KEY,
+			( builder, path ) -> builder.defineEnum( path, GenerationType.SERIAL )
+		);
+		registerConfigValue(
+			"How many chunk generation tasks shall be start in parallel? " +
+				"If the value is \"0\", the number of processor cores is used.",
+			GENERATION_SEMI_PARALLEL_TASK_COUNT_KEY,
+			( builder, path ) -> builder.defineInRange( path, 0, 0, getProcessorCount() << 1 )
+		);
+		pop();
 		push( "Delays of the print and save tasks", DELAYS_KEY );
 		registerConfigValue(
 			"Time between 2 status prints",
@@ -64,15 +85,6 @@ public class ServerConfig extends AbstractConfig {
 			"Time between 2 saving events",
 			SAVE_DELAY_KEY,
 			( builder, path ) -> builder.defineInRange( path, 180, 1, Integer.MAX_VALUE )
-		);
-		pop();
-		push( "Parameters for the parallel generation of several chunks", PARALLEL_KEY );
-		registerConfigValue( "Should the chunks be generated in parallel?", PARALLEL_ENABLED_KEY, false );
-		registerConfigValue(
-			"How many chunk generation shall run in parallel? " +
-				"If the value is \"0\", the number of processor cores is used.",
-			PARALLEL_COUNT_KEY,
-			( builder, path ) -> builder.defineInRange( path, 0, 0, getProcessorCount() << 1 )
 		);
 		pop();
 	}
@@ -97,14 +109,15 @@ public class ServerConfig extends AbstractConfig {
 		return getValue( Integer.class, SAVE_DELAY_KEY );
 	}
 	
-	public boolean isRunParallel() {
+	@NotNull
+	public GenerationType getGenerationType() {
 		
-		return getValue( Boolean.class, PARALLEL_ENABLED_KEY );
+		return getValue( GenerationType.class, GENERATION_TYPE_KEY );
 	}
 	
-	public int getThreadCount() {
+	public int getGenerationSemiParallelTaskCount() {
 		
-		int threadCount = getValue( Integer.class, PARALLEL_COUNT_KEY );
+		int threadCount = getValue( Integer.class, GENERATION_SEMI_PARALLEL_TASK_COUNT_KEY );
 		return threadCount == 0 ? getProcessorCount() : threadCount;
 	}