+++ src/main/java/de/geheimagentnr1/world_pre_generator/elements/commands/PregenCommand.java
@@ -8,7 +8,10 @@ import com.mojang.brigadier.context.CommandContext;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import de.geheimagentnr1.minecraft_forge_api.elements.commands.CommandInterface;
 import de.geheimagentnr1.world_pre_generator.config.ServerConfig;
+import de.geheimagentnr1.world_pre_generator.elements.commands.models.TaskRow;
+import de.geheimagentnr1.world_pre_generator.elements.commands.models.TasksMessage;
 import de.geheimagentnr1.world_pre_generator.elements.queues.tasks.pregen.PregenTask;
+import de.geheimagentnr1.world_pre_generator.elements.queues.tasks.pregen.TaskType;
 import de.geheimagentnr1.world_pre_generator.elements.queues.tasks.pregen.data.WorldPos;
 import de.geheimagentnr1.world_pre_generator.elements.workers.PregenWorker;
 import de.geheimagentnr1.world_pre_generator.helpers.DimensionHelper;
@@ -17,9 +20,11 @@ import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.commands.Commands;
 import net.minecraft.commands.arguments.DimensionArgument;
 import net.minecraft.commands.arguments.coordinates.ColumnPosArgument;
+import net.minecraft.core.SectionPos;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.server.level.ColumnPos;
+import net.minecraft.world.level.ChunkPos;
 import net.minecraft.world.level.Level;
 import org.jetbrains.annotations.NotNull;
 
@@ -45,6 +50,17 @@ public class PregenCommand implements CommandInterface {
 			.requires( source -> source.hasPermission( 2 ) );
 		pregenCommand.then( Commands.literal( "clear" )
 			.executes( this::clear ) );
+		
+		LiteralArgumentBuilder<CommandSourceStack> startCommand = Commands.literal( "start" );
+		for( TaskType taskType : TaskType.values() ) {
+			startCommand.then( Commands.literal( taskType.getSerializedName() )
+				.then( Commands.argument( "center", ColumnPosArgument.columnPos() )
+					.then( Commands.argument( "radius", IntegerArgumentType.integer( 1 ) )
+						.executes( context -> start( context, taskType, false ) )
+						.then( Commands.argument( "force", BoolArgumentType.bool() )
+							.executes( context -> start( context, taskType, true ) ) ) ) ) );
+		}
+		
 		pregenCommand.then( Commands.literal( "gen" )
 			.then( Commands.argument( "dimension", DimensionArgument.dimension() )
 				.then( Commands.literal( "cancel" )
@@ -53,12 +69,7 @@ public class PregenCommand implements CommandInterface {
 					.executes( this::pause ) )
 				.then( Commands.literal( "resume" )
 					.executes( this::resume ) )
-				.then( Commands.literal( "start" )
-					.then( Commands.argument( "center", ColumnPosArgument.columnPos() )
-						.then( Commands.argument( "radius", IntegerArgumentType.integer( 1 ) )
-							.executes( this::startUnforced )
-							.then( Commands.argument( "force", BoolArgumentType.bool() )
-								.executes( this::start ) ) ) ) ) ) );
+				.then( startCommand ) ) );
 		pregenCommand.then( Commands.literal( "list" )
 			.executes( this::printList ) );
 		pregenCommand.then( Commands.literal( "sendFeedback" )
@@ -66,22 +77,6 @@ public class PregenCommand implements CommandInterface {
 			.then( Commands.argument( "isFeedbackEnabled", BoolArgumentType.bool() )
 				.executes( this::setSendFeedback ) ) );
 		return pregenCommand;
-	}
-	
-	private void printTasks( @NotNull CommandSourceStack source, @NotNull ArrayList<PregenTask> tasks ) {
-		
-		for( PregenTask task : tasks ) {
-			source.sendSuccess(
-				() -> Component.literal( String.format(
-					"%s %d %d %d",
-					DimensionHelper.getNameOfDim( task.getDimension() ),
-					task.getCenterX(),
-					task.getCenterZ(),
-					task.getRadius()
-				) ),
-				false
-			);
-		}
 	}
 	
 	private int printList( @NotNull CommandContext<CommandSourceStack> context ) {
@@ -105,27 +100,52 @@ public class PregenCommand implements CommandInterface {
 		return Command.SINGLE_SUCCESS;
 	}
 	
-	private int startUnforced( @NotNull CommandContext<CommandSourceStack> context ) throws CommandSyntaxException {
+	private void printTasks( @NotNull CommandSourceStack source, @NotNull ArrayList<PregenTask> tasks ) {
 		
-		return start( context, false );
+		TasksMessage tasksMessage = new TasksMessage();
+		tasksMessage.add( new TaskRow( "Dimension", "Type", "Center", null, null, "Radius", "Forced" ) );
+		tasksMessage.addAll( tasks );
+		for( String message : tasksMessage.buildMessages() ) {
+			source.sendSuccess(
+				() -> Component.literal( message ),
+				false
+			);
+		}
 	}
 	
-	private int start( @NotNull CommandContext<CommandSourceStack> context ) throws CommandSyntaxException {
-		
-		return start( context, BoolArgumentType.getBool( context, "force" ) );
-	}
-	
-	private int start( @NotNull CommandContext<CommandSourceStack> context, boolean force )
+	private int start(
+		@NotNull CommandContext<CommandSourceStack> context,
+		@NotNull TaskType taskType,
+		boolean checkForced )
 		throws CommandSyntaxException {
 		
 		CommandSourceStack source = context.getSource();
 		ColumnPos center = ColumnPosArgument.getColumnPos( context, "center" );
 		int radius = IntegerArgumentType.getInteger( context, "radius" );
+		int chunkX;
+		int chunkZ;
+		int chunkRadius;
+		switch( taskType ) {
+			case BLOCK -> {
+				ChunkPos chunkCenter = center.toChunkPos();
+				chunkX = chunkCenter.x;
+				chunkZ = chunkCenter.z;
+				chunkRadius = SectionPos.blockToSectionCoord( radius );
+			}
+			case CHUNK -> {
+				chunkX = center.x();
+				chunkZ = center.z();
+				chunkRadius = radius;
+			}
+			default -> throw new IllegalArgumentException( "Invalid task type" );
+		}
 		ResourceKey<Level> dimension = DimensionArgument.getDimension( context, "dimension" ).dimension();
+		boolean force = checkForced && BoolArgumentType.getBool( context, "force" );
 		
 		pregenWorker.getQueue().startTask( new PregenTask(
-			new WorldPos( center.toChunkPos().x, center.toChunkPos().z ),
-			radius,
+			taskType,
+			new WorldPos( chunkX, chunkZ ),
+			chunkRadius,
 			dimension,
 			force
 		) );