+++ src/main/java/de/geheimagentnr1/world_pre_generator/elements/queues/tasks/pregen/TaskType.java
@@ -1,0 +1,10 @@
+package de.geheimagentnr1.world_pre_generator.elements.queues.tasks.pregen;
+
+import de.geheimagentnr1.minecraft_forge_api.util.SimpleStringRepresentable;
+
+
+public enum TaskType implements SimpleStringRepresentable {
+	
+	BLOCK,
+	CHUNK
+}